refactor: move files to correct folder, extract functions into utils
