+++ packages/shared/lib/core/wallet/actions/filterQueriedActivities.ts
@@ -1,0 +1,14 @@
+import { ActivityDirection } from '@core/wallet'
+import { get } from 'svelte/store'
+import { activities, queriedActivities } from '../stores'
+
+export function filterQueriedActivities(filter: number): void {
+    let activityList = get(activities)
+
+    if (filter === 1) {
+        activityList = activityList.filter((activity) => activity.direction === ActivityDirection.In)
+    } else if (filter === 2) {
+        activityList = activityList.filter((activity) => activity.direction === ActivityDirection.Out)
+    }
+    queriedActivities.set(activityList)
+}
