+++ packages/shared/lib/core/wallet/actions/index.ts
@@ -1,0 +1,2 @@
+export * from './filterQueriedActivities'
+export * from './searchQueriedActivities'
