+++ packages/shared/lib/core/wallet/interfaces/index.ts
@@ -1,3 +1,4 @@
+export * from './activity.interface'
 export * from './asset-balance.interface'
 export * from './asset.interface'
 export * from './token-metadata.interface'
