+++ packages/shared/lib/core/wallet/stores/activities.store.ts
@@ -1,11 +1,12 @@
 import { selectedAccount } from '@core/account'
-import { derived, get, Readable, writable, Writable } from 'svelte/store'
+import { derived, Readable, writable, Writable } from 'svelte/store'
 import { getAccountMessages } from '@lib/wallet'
 import { parseTransactionsToActivities } from '../utils'
-import { ActivityDirection, IActivity } from '@lib/typings/activity'
-import { isValueInUnitRange, unitToValue } from '@lib/utils'
-import { formatUnitBestMatch } from 'shared/lib/units'
-import { formatDate, localize } from '@core/i18n'
+import { getMonthYear } from '@lib/utils'
+import { Transaction } from '@lib/typings/message'
+import { AccountMessage } from '@lib/typings/wallet'
+import { localize } from '@core/i18n'
+import { IActivity } from '../interfaces/activity.interface'
 
 export const activities: Readable<IActivity[]> = derived([selectedAccount], ([$selectedAccount]) =>
     parseTransactionsToActivities(getAccountMessages($selectedAccount))
@@ -31,40 +32,7 @@ export const groupedActivities: Readable<GroupedActivity[]> = derived([queriedAc
     return groupedActivities
 })
 
-export function searchQueriedActivities(searchTerm: string): void {
-    queriedActivities.set(
-        get(activities).filter((activity) => {
-            const amount = unitToValue(activity.amount)
-            return (
-                activity.subjectAccountName === searchTerm ||
-                activity.subjectAddress === searchTerm ||
-                activity?.id.toLowerCase() === searchTerm ||
-                (searchTerm[0] === '>' && unitToValue(searchTerm.substring(1)) < amount) ||
-                (searchTerm[0] === '<' && unitToValue(searchTerm.substring(1)) > amount) ||
-                (searchTerm[1] === 'i' && isValueInUnitRange(amount, searchTerm)) ||
-                amount === unitToValue(searchTerm) ||
-                formatUnitBestMatch(amount).toString().toLowerCase()?.includes(searchTerm)
-            )
-        })
-    )
-}
-
-export function filterQueriedActivities(filter: number): void {
-    let activityList = get(activities)
-
-    if (filter === 1) {
-        activityList = activityList.filter((activity) => activity.direction === ActivityDirection.In)
-    } else if (filter === 2) {
-        activityList = activityList.filter((activity) => activity.direction === ActivityDirection.Out)
-    }
-    queriedActivities.set(activityList)
-}
-
 function getDateString(timestamp): string {
     const dateString = getMonthYear(new Date(Number(timestamp)))
     return dateString === getMonthYear(new Date()) ? localize('general.thisMonth') : dateString
-}
-
-function getMonthYear(date): string {
-    return formatDate(date, { year: 'numeric', month: 'short' })
 }
