+++ packages/shared/lib/core/wallet/stores/index.ts
@@ -1,2 +1,2 @@
-export * from './assets.store'
 export * from './activities.store'
+export * from './assets.store'
