+++ packages/shared/lib/typings/index.ts
@@ -1,5 +1,4 @@
 import * as account from './account'
-import * as activity from './activity'
 import * as address from './address'
 import * as app from './app'
 import * as bridge from './bridge'
